fix(web): move server-only helper out of client-importable module
Splits `readActiveLegalEntityId` out of `get-active-legal-entity.ts`
into `src/server/active-legal-entity.server.ts`. The `*.server.ts`
convention is recognised by TanStack Start's Vite plugin (default
`server.files` rule: `**/*.server.*`), which prevents the file from
being included in the client bundle and eliminates the
`dotenv/config` → `process.argv.reduce` crash in the browser.
